Get types right

+++ src/hardware/gpu/gpu_comparison_kernels.cu
@@ -15,7 +15,7 @@
 #include <cuda.h>
 
 // GPU: Truth value testing
-__global__ void gpu_logical_all(float *A, int size, bool &result){
+__global__ void gpu_logical_all(float *A, long int size, bool &result){
     long int thread_id_x = blockIdx.x * blockDim.x + threadIdx.x;
 
     // if(!result) return;  // Abort if there is a result
@@ -28,7 +28,7 @@ __global__ void gpu_logical_all(float *A, int size, bool &result){
     }
 }
 
-__global__ void gpu_logical_any(float *A, int size, bool &result){
+__global__ void gpu_logical_any(float *A, long int size, bool &result){
     long int thread_id_x = blockIdx.x * blockDim.x + threadIdx.x;
 
     // if(result) return;  // Abort if there is a result
@@ -41,7 +41,7 @@ __global__ void gpu_logical_any(float *A, int size, bool &result){
     }
 }
 
-__global__ void gpu_isfinite(float *A, float *B, int size){
+__global__ void gpu_isfinite(float *A, float *B, long int size){
     long int thread_id_x = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (thread_id_x < size){
@@ -49,7 +49,7 @@ __global__ void gpu_isfinite(float *A, float *B, int size){
     }
 }
 
-__global__ void gpu_isinf(float *A, float *B, int size){
+__global__ void gpu_isinf(float *A, float *B, long int size){
     long int thread_id_x = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (thread_id_x < size){
@@ -57,7 +57,7 @@ __global__ void gpu_isinf(float *A, float *B, int size){
     }
 }
 
-__global__ void gpu_isnan(float *A, float *B, int size){
+__global__ void gpu_isnan(float *A, float *B, long int size){
     long int thread_id_x = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (thread_id_x < size){
@@ -65,7 +65,7 @@ __global__ void gpu_isnan(float *A, float *B, int size){
     }
 }
 
-__global__ void gpu_isneginf(float *A, float *B, int size){
+__global__ void gpu_isneginf(float *A, float *B, long int size){
     long int thread_id_x = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (thread_id_x < size){
@@ -73,7 +73,7 @@ __global__ void gpu_isneginf(float *A, float *B, int size){
     }
 }
 
-__global__ void gpu_isposinf(float *A, float *B, int size){
+__global__ void gpu_isposinf(float *A, float *B, long int size){
     long int thread_id_x = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (thread_id_x < size){
@@ -82,7 +82,7 @@ __global__ void gpu_isposinf(float *A, float *B, int size){
 }
 
 
-__global__ void gpu_logical_and(float *A, float *B, float *C, int size){
+__global__ void gpu_logical_and(float *A, float *B, float *C, long int size){
     long int thread_id_x = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (thread_id_x < size){
@@ -90,7 +90,7 @@ __global__ void gpu_logical_and(float *A, float *B, float *C, int size){
     }
 }
 
-__global__ void gpu_logical_or(float *A, float *B, float *C, int size){
+__global__ void gpu_logical_or(float *A, float *B, float *C, long int size){
     long int thread_id_x = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (thread_id_x < size){
@@ -98,7 +98,7 @@ __global__ void gpu_logical_or(float *A, float *B, float *C, int size){
     }
 }
 
-__global__ void gpu_logical_not(float *A, float *B, int size){
+__global__ void gpu_logical_not(float *A, float *B, long int size){
     long int thread_id_x = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (thread_id_x < size){
@@ -106,7 +106,7 @@ __global__ void gpu_logical_not(float *A, float *B, int size){
     }
 }
 
-__global__ void gpu_logical_xor(float *A, float *B, float *C, int size){
+__global__ void gpu_logical_xor(float *A, float *B, float *C, long int size){
     long int thread_id_x = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (thread_id_x < size){
@@ -115,7 +115,7 @@ __global__ void gpu_logical_xor(float *A, float *B, float *C, int size){
 }
 
 
-__global__  void gpu_logical_allclose(float *A, float *B, float rtol, float atol, bool equal_nan, int size, bool &allclose){
+__global__  void gpu_logical_allclose(float *A, float *B, float rtol, float atol, bool equal_nan, long int size, bool &allclose){
     long int thread_id_x = blockIdx.x * blockDim.x + threadIdx.x;
 
     // if(!allclose) return;  // Abort if there is a result
@@ -129,7 +129,7 @@ __global__  void gpu_logical_allclose(float *A, float *B, float rtol, float atol
     }
 }
 
-__global__  void gpu_logical_isclose(float *A, float *B, float *C, float rtol, float atol, bool equal_nan, int size){
+__global__  void gpu_logical_isclose(float *A, float *B, float *C, float rtol, float atol, bool equal_nan, long int size){
     long int thread_id_x = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (thread_id_x < size){
@@ -137,7 +137,7 @@ __global__  void gpu_logical_isclose(float *A, float *B, float *C, float rtol, f
     }
 }
 
-__global__ void gpu_greater(float *A, float *B, float v, int size){
+__global__ void gpu_greater(float *A, float *B, float v, long int size){
     long int thread_id_x = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (thread_id_x < size){
@@ -145,7 +145,7 @@ __global__ void gpu_greater(float *A, float *B, float v, int size){
     }
 }
 
-__global__ void gpu_greater(float *A, float *B, float *C, int size){
+__global__ void gpu_greater(float *A, float *B, float *C, long int size){
     long int thread_id_x = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (thread_id_x < size){
@@ -153,7 +153,7 @@ __global__ void gpu_greater(float *A, float *B, float *C, int size){
     }
 }
 
-__global__  void gpu_greater_equal(float *A, float *B, float v, int size){
+__global__  void gpu_greater_equal(float *A, float *B, float v, long int size){
     long int thread_id_x = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (thread_id_x < size){
@@ -161,7 +161,7 @@ __global__  void gpu_greater_equal(float *A, float *B, float v, int size){
     }
 }
 
-__global__  void gpu_greater_equal(float *A, float *B, float *C, int size){
+__global__  void gpu_greater_equal(float *A, float *B, float *C, long int size){
     long int thread_id_x = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (thread_id_x < size){
@@ -169,7 +169,7 @@ __global__  void gpu_greater_equal(float *A, float *B, float *C, int size){
     }
 }
 
-__global__  void gpu_less(float *A, float *B, float v, int size){
+__global__  void gpu_less(float *A, float *B, float v, long int size){
     long int thread_id_x = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (thread_id_x < size){
@@ -177,7 +177,7 @@ __global__  void gpu_less(float *A, float *B, float v, int size){
     }
 }
 
-__global__  void gpu_less(float *A, float *B, float *C, int size){
+__global__  void gpu_less(float *A, float *B, float *C, long int size){
     long int thread_id_x = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (thread_id_x < size){
@@ -185,7 +185,7 @@ __global__  void gpu_less(float *A, float *B, float *C, int size){
     }
 }
 
-__global__  void gpu_less_equal(float *A, float *B, float v, int size){
+__global__  void gpu_less_equal(float *A, float *B, float v, long int size){
     long int thread_id_x = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (thread_id_x < size){
@@ -193,7 +193,7 @@ __global__  void gpu_less_equal(float *A, float *B, float v, int size){
     }
 }
 
-__global__  void gpu_less_equal(float *A, float *B, float *C, int size){
+__global__  void gpu_less_equal(float *A, float *B, float *C, long int size){
     long int thread_id_x = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (thread_id_x < size){
@@ -201,7 +201,7 @@ __global__  void gpu_less_equal(float *A, float *B, float *C, int size){
     }
 }
 
-__global__  void gpu_equal(float *A, float *B, float v, int size){
+__global__  void gpu_equal(float *A, float *B, float v, long int size){
     long int thread_id_x = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (thread_id_x < size){
@@ -209,7 +209,7 @@ __global__  void gpu_equal(float *A, float *B, float v, int size){
     }
 }
 
-__global__  void gpu_equal(float *A, float *B, float *C, int size){
+__global__  void gpu_equal(float *A, float *B, float *C, long int size){
     long int thread_id_x = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (thread_id_x < size){
@@ -218,7 +218,7 @@ __global__  void gpu_equal(float *A, float *B, float *C, int size){
 }
 
 
-__global__  void gpu_not_equal(float *A, float *B, float v, int size){
+__global__  void gpu_not_equal(float *A, float *B, float v, long int size){
     long int thread_id_x = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (thread_id_x < size){
@@ -226,7 +226,7 @@ __global__  void gpu_not_equal(float *A, float *B, float v, int size){
     }
 }
 
-__global__  void gpu_not_equal(float *A, float *B, float *C, int size){
+__global__  void gpu_not_equal(float *A, float *B, float *C, long int size){
     long int thread_id_x = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (thread_id_x < size){
